feat(goexports): support multi os / arch, generate syscalls for all platforms (#6)

+++ stdlib/hash_crc32.go
@@ -9,18 +9,19 @@ import (
 
 func init() {
 	Value["hash/crc32"] = map[string]reflect.Value{
-		"Castagnoli":   reflect.ValueOf(crc32.Castagnoli),
+		"Castagnoli":   reflect.ValueOf(uint32(crc32.Castagnoli)),
 		"Checksum":     reflect.ValueOf(crc32.Checksum),
 		"ChecksumIEEE": reflect.ValueOf(crc32.ChecksumIEEE),
-		"IEEE":         reflect.ValueOf(crc32.IEEE),
+		"IEEE":         reflect.ValueOf(uint32(crc32.IEEE)),
 		"IEEETable":    reflect.ValueOf(crc32.IEEETable),
-		"Koopman":      reflect.ValueOf(crc32.Koopman),
+		"Koopman":      reflect.ValueOf(uint32(crc32.Koopman)),
 		"MakeTable":    reflect.ValueOf(crc32.MakeTable),
 		"New":          reflect.ValueOf(crc32.New),
 		"NewIEEE":      reflect.ValueOf(crc32.NewIEEE),
 		"Size":         reflect.ValueOf(crc32.Size),
 		"Update":       reflect.ValueOf(crc32.Update),
 	}
+
 	Type["hash/crc32"] = map[string]reflect.Type{
 		"Table": reflect.TypeOf((*crc32.Table)(nil)).Elem(),
 	}
