+++ stdlib/hash_crc64.go
@@ -10,13 +10,14 @@ import (
 func init() {
 	Value["hash/crc64"] = map[string]reflect.Value{
 		"Checksum":  reflect.ValueOf(crc64.Checksum),
-		"ECMA":      reflect.ValueOf(uint(crc64.ECMA)),
-		"ISO":       reflect.ValueOf(uint(crc64.ISO)),
+		"ECMA":      reflect.ValueOf(uint64(crc64.ECMA)),
+		"ISO":       reflect.ValueOf(uint64(crc64.ISO)),
 		"MakeTable": reflect.ValueOf(crc64.MakeTable),
 		"New":       reflect.ValueOf(crc64.New),
 		"Size":      reflect.ValueOf(crc64.Size),
 		"Update":    reflect.ValueOf(crc64.Update),
 	}
+
 	Type["hash/crc64"] = map[string]reflect.Type{
 		"Table": reflect.TypeOf((*crc64.Table)(nil)).Elem(),
 	}
